install headers consistently in subfolder
Without this patch catkin will set the correct include folder for devel workspaces only (apriltags2_ros/apriltags2/include/);
using the package with an install workspace fails.
The CMakeLists.txt installs the package headers to .../include/apriltags2/ (this is a good official location
to install to because without the subfolder folder the individual headers would clutter .../include).
But catkin automatically includes only .../include when the installed package is find_packaged.

With this change the official way to include the headers from the apriltags2 package becomes
`#include <apriltags2/apriltag.h>` which is consistent between devel and install workspace.